requirements.txt
alembic
dotenv
faiss-cpu
fastapi
google-generativeai
kaggle
uvicorn
pandas
pinecone
pymongo
psycopg2-binary
pytest
python-multipart
requests
sentence-transformers
sqlalchemy
streamlit
uvicorn